models: Add method to convert ScheduleMessage objects into dicts.
